pyproject.toml
[project]
name = "workflow-agent"
version = "0.1.0"
description = "A Python framework for orchestrating multi-step workflows with AI-driven adaptation."
authors = [
  { name = "Your Name", email = "you@example.com" }
]
dependencies = [
  "jinja2==3.1.2",
  "typer==0.9.0",
  "rich==13.3.2",
  "InquirerPy==0.3.4",
  "python-dotenv==1.0.0",
  "PyYAML==6.0",
  "SQLAlchemy==1.4.47",
  "psutil==5.9.5",
  "pydantic==1.10.7",
  "langchain-openai==0.0.1",  # Optional: required only if using LLM optimization.
  "shellcheck-py==0.9.0",     # Added: Alternative to LLM for script validation
  "asyncio==3.4.3",           # Added: For concurrent execution
  "aiosqlite==0.17.0",        # Added: For async database operations
  "aiohttp==3.8.5",           # Added: For async HTTP requests
  "pytest==7.3.1",            # Added: For testing
  "pytest-asyncio==0.21.0"    # Added: For testing async code
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"